Backend/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
python-dotenv
faster-whisper
sentence-transformers
